% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicInput.R
\name{dynamicInput}
\alias{dynamicInput}
\title{Page Template for Keboola Shiny Apps
Dynamic Inputs
This element is like a multi-select, but for each selection 
a range element is created for the selected column}
\usage{
dynamicInput(inputId, label, choices = c(), multiple = TRUE, ...)
}
\arguments{
\item{inputId}{character identifier of the element}

\item{label}{character label}
}
\description{
Page Template for Keboola Shiny Apps
Dynamic Inputs
This element is like a multi-select, but for each selection 
a range element is created for the selected column
}

