% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keboolaModal.R
\name{keboolaModalButton}
\alias{keboolaModalButton}
\title{Button to create modal dialogs for Keboola Shiny Apps}
\usage{
keboolaModalButton(inputId, label, icon = NULL, title = "", content = "")
}
\arguments{
\item{inputId}{-- input ID for the modal window}

\item{label}{-- Dialog title}

\item{icon}{-- optional dialog icon}

\item{content}{-- dialog content (can include input and output elements)}

\item{header}{-- content for dialog header}
}
\description{
Button to create modal dialogs for Keboola Shiny Apps
}

