% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keboolaPage.R
\name{keboolaPage}
\alias{keboolaPage}
\title{Page Template for Keboola Shiny Apps}
\usage{
keboolaPage(bootstrapPage)
}
\arguments{
\item{bootstrapPage}{Bootstrape page element}
}
\description{
Page Template for Keboola Shiny Apps
}

