% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keboola-shiny.R
\docType{class}
\name{KeboolaShiny-class}
\alias{KeboolaShiny}
\alias{KeboolaShiny-class}
\title{Helper Library for Keboola Shiny Applications}
\description{
Helper Library for Keboola Shiny Applications
}

