% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeboolaAppConfig.R
\docType{class}
\name{KeboolaAppConfig-class}
\alias{KeboolaAppConfig}
\alias{KeboolaAppConfig-class}
\title{This is the configuration component library for managing app configurations}
\arguments{
\item{sapiClient}{- Keboola.sapi.r.client::SapiClient}

\item{bucket}{- Bucket where config table is stored}

\item{shinyUrl}{- Shiny Bundle API home URL
it will be read from command line argument.
"http://shiny.kbc-devel-02.keboola.com/app_dev.php/shiny/"}
}
\description{
This is the configuration component library for managing app configurations
}
\section{Methods}{

\describe{
\item{\code{clearForm(input)}}{TODO
\subsection{Parameters}{\itemize{
\item{\code{input} TODO}
}}
\subsection{Return Value}{TODO}}

\item{\code{configChoices()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{configListUI()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{configs()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{configSettingsUI()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{deleteConfig(configId)}}{TODO
\subsection{Parameters}{\itemize{
\item{\code{configId} TODO}
}}
\subsection{Return Value}{TODO}}

\item{\code{deleteConfigResultUI()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{getConfigs()}}{Get app configurations from the Shiny Bundle.
\subsection{Return Value}{List of app configurations.}}

\item{\code{loadConfigResultUI(callback)}}{TODO
\subsection{Parameters}{\itemize{
\item{\code{callback} TODO}
}}
\subsection{Return Value}{TODO}}

\item{\code{saveConfig()}}{This method stores the entire session$input object in the kbc storage
in the bucket being used by the app
\subsection{Return Value}{TRUE, will throw an error if something goes wrong.}}

\item{\code{saveConfigResultUI()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{saveConfigUI()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{selectedConfig()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{settingsModalButton()}}{TODO
\subsection{Return Value}{TODO}}
}}

