% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KeboolaAppConfig.R
\docType{class}
\name{KeboolaAppConfig-class}
\alias{KeboolaAppConfig}
\alias{KeboolaAppConfig-class}
\title{This is the configuration component library for managing app configurations}
\description{
This is the configuration component library for managing app configurations
}

