% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KeboolaAppData.R
\docType{class}
\name{KeboolaAppData-class}
\alias{KeboolaAppData}
\alias{KeboolaAppData-class}
\title{This is the configuration component library for managing app data loading / saving}
\description{
This is the configuration component library for managing app data loading / saving
}

