% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keboolaPage.R
\name{keboolaPage}
\alias{keboolaPage}
\title{Page Template for Keboola Shiny Apps}
\usage{
keboolaPage(sidebarPanel, mainPanel)
}
\arguments{
\item{sidebarPanel}{UI element}

\item{mainPanel}{UI element}
}
\description{
Page Template for Keboola Shiny Apps
}

