% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keboolaPage.R
\name{keboolaPage}
\alias{keboolaPage}
\title{Page Template for Keboola Shiny Apps}
\usage{
keboolaPage(page, appTitle = "Default")
}
\arguments{
\item{appTitle}{Application title}

\item{bootstrapPage}{bootstrapePage element}
}
\description{
Page Template for Keboola Shiny Apps
}

