% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keboola-shiny.R
\docType{class}
\name{KeboolaShiny-class}
\alias{KeboolaShiny}
\alias{KeboolaShiny-class}
\title{Helper Library for Keboola Shiny Applications}
\description{
Helper Library for Keboola Shiny Applications
}
\section{Fields}{

\describe{
\item{\code{session}}{Shiny session object}

\item{\code{loggedIn}}{1 if the user is logged in, 0 if not}

\item{\code{errMsg}}{Error encountered when logging in}

\item{\code{loginErrrorOutput}}{HTML element with login error}

\item{\code{bucketId}}{ID of the current bucket}

\item{\code{runId}}{ID of the current job}

\item{\code{token}}{Current KBC token}

\item{\code{appId}}{TODO}

\item{\code{client}}{Instance of Storage API client}

\item{\code{db}}{Instance of Redshift Driver}

\item{\code{kfig}}{Instance of Application Config}

\item{\code{kdata}}{Instance of Application Data}

\item{\code{userRunId}}{TODO}

\item{\code{loading}}{TODO}
}}
\section{Methods}{

\describe{
\item{\code{concludeStartup(options)}}{TODO
\subsection{Parameters}{\itemize{
\item{\code{options} TODO}
}}
\subsection{Return Value}{TODO}}

\item{\code{dbConnect()}}{Establish a connection via provisioning client credentials.
\subsection{Return Value}{TRUE}}

\item{\code{getAppId()}}{Get the AppId from the URL.
\subsection{Return Value}{Application ID of this registered app}}

\item{\code{getBucket()}}{Get the bucket from the query string.
\subsection{Return Value}{String storage Bucket ID.}}

\item{\code{getBuckets()}}{Return a list of accessible buckets.
\subsection{Return Value}{Character vector of bucketI IDs}}

\item{\code{getConfigId()}}{Get the configId from the URL.
\subsection{Return Value}{}}

\item{\code{getLogin()}}{Verify login information and initialize application.
\subsection{Return Value}{List with items:
token, runId, bucket, loggedIn, errMsg, client, ready
}}

\item{\code{getRunId()}}{Get the runId from the query string.
\subsection{Return Value}{String Run ID.}}

\item{\code{getToken()}}{Get the token from the session headers or input element.
\subsection{Return Value}{String KBC token}}

\item{\code{initialize(requireRunId = TRUE, session = getDefaultReactiveDomain())}}{Constructor.
\subsection{Parameters}{\itemize{
\item{\code{requireRunId} Use RunId.}
\item{\code{session} Shiny session object.}
}}}

\item{\code{initLibs()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{loadTable(prettyName, name)}}{Load table from Storage. 
(Exposed wrapper for \code{KeboolaAppData} method).
\subsection{Parameters}{\itemize{
\item{\code{prettyName} Table name to be used in labels throughout the app.}
\item{\code{name} Name of table in SAPI.}
}}
\subsection{Return Value}{data.frame with table data}}

\item{\code{loadTables(tables, options)}}{TODO
\subsection{Parameters}{\itemize{
\item{\code{tables} TODO}
\item{\code{options} TODO}
}}
\subsection{Return Value}{TODO}}

\item{\code{ready()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{sourceData()}}{TODO
\subsection{Return Value}{TODO}}

\item{\code{startup(options = list(appTitle = "", tables = list(), cleanData = FALSE,
  dataToSave = NULL, configCallback = NULL, description = FALSE, customElements
  = NULL))}}{This method returns a list of all elements from the shared library that are destined for the shiny server output object
They are mainly renderUI functions, but can be any render function available in shiny
\subsection{Parameters}{\itemize{
\item{\code{options} List with items:
appTitle - the title of the application
tables - list of tables to load from sapi list(localname = sapiname)
dataToSave - the reactive in server.R that holds the input filtered data, or any data that would want to be saved to sapi
configCallback - the method which is to be invoked when a configuration is loaded.  This method will generally update inputs according to the selectedConfig values
description - whether or not to include a description object
customElements - the method for processing custom elements of the description}
}}
\subsection{Return Value}{TODO}}
}}

